Allows GamePlay to be compiled on non-pvr systems
Modifies 'texture' to optionally compile in the PVR compressed
texture defines and functions if the platform supports PVR

Currently this check is probably insufficient (do all Android devices
use a PowerVR graphics chip?)
The #ifdef's in Texture are also not very nice, especially because
two of them are in the header file.

Should probably be extracted to simply 'compressTexture' and then
the texture object compresses the texture with PVR on those platforms
that support it and with other tools (E.g. DXT1 or S3TC). I am not
sure however, how this works with differeing graphics cards such as
nvida vs. AMD ATI.