Modified how __orientationAngle is calculated on Android.
The assumption that all android devices default orientation is portrait is incorrect. Tablets such as Transformer prime have the default orientation as Landscape. This leads to the axis of the tilt sensors being incorrect.

I have changed the androidPlatform so that it uses Display.getRotation in the android API so that the tilt sensors are correct.