Fixes unsupported framebuffer issue with  qualcomm's android driver.
When DepthStencilTargets are created using GL_DEPTH24_STENCIL8_OES
the same render buffer has to be used as depth and stencil attachment.

It seems that other driver implementations are not as restrictive as
qualcomm's and allow different renderbuffers. On tested Qualcomm Adreno
200 and 320 android devices however this resulted in an unsupported
Framebuffer.

Issue is fixed by setting a flag whether or not the DepthStencilTarget
uses an packed format. This flag is considered when attaching renderbuffers.